requirements.txt
tensorflow==2.15.0
fastapi==0.110.0
uvicorn[standard]==0.27.1
pillow==10.2.0
numpy==1.26.4
pandas==2.2.0
python-multipart==0.0.9
scikit-learn==1.4.0
tqdm==4.66.2
matplotlib==3.8.2
seaborn==0.13.2
opencv-python==4.9.0.80
streamlit